Restructure production and dev environment, remove db from production api compose

+++ scripts/docker/remove_images.sh
@@ -4,39 +4,46 @@
 # First it finds all related containers, stops them and removes them.
 # After removing containers, images can be safely stopped.
 
-images=$(docker images -a | awk '{ print $1,$3 }')
-
-echo $images
-
-exit 0
+images=$(docker images -a | awk '{if(NR>1)print}' | awk '{ print $3 }')
 
 if [ "$images" == "" ]; then
     echo "No images found."
     exit 0
 fi
 
-
-for image in $images; do
-   
-    echo "Image: $image"
-
-    # Find containers usibng image
-    containers=$(docker ps -a | awk '{ print $1,$2 }' | grep $image)
+for image_id in $images; do
     
-    if [ "$containers" == "" ]; then
-        echo "No containers to stop"
+    image_name=$(docker images -a | awk '{if(NR>1)print}' | grep $image_id | awk '{ print $1 }')
+    
+    if [ "$image_name" = "<none>" ]; then
         continue
     fi
+
     
-    echo "Used by containers:"
+    if [ "$image_name" = "mysql" ]; then
+        echo "Skipping mysql image and not removing any containers associated with it"
+        continue
+    fi
+
+    echo "Image: $image_name id: $image_id"
     
-    # Stop & remove all found containers
-    for container in $containers; do
-        echo "-> $container"
-     	docker container stop $container
-	    docker container rm $container
-    done
+    # Find containers using image
+    containers=$(docker ps -a | awk '{ print $1,$2 }' | grep $image_id)
     
-    # Remove image
-    docker rmi $image
+    if [ "$containers" == "" ]; then
+        docker rmi $image_id
+        continue
+    else
+        echo "Used by containers:"
+    
+        # Stop & remove all found containers
+        for container in $containers; do
+            echo "-> $container"
+     	    docker container stop $container
+	        docker container rm $container
+        done
+    fi
+
 done
+
+echo "Removed all images and their containers (skipped mysql)".
